jouer le son meme si on sort de cette page
