+++ lib/screens/music_screen.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:tiger_vibes/models/songs-models.dart';
 import 'package:tiger_vibes/screens/musicPlayer_page.dart';
-import '../music_service.dart'; // Assurez-vous que le chemin d'importation est correct
-
+import '../music_service.dart'; 
 class MusicListPage extends StatefulWidget {
   const MusicListPage({super.key});
 
@@ -16,7 +15,7 @@ class _MusicListPageState extends State<MusicListPage> {
   @override
   void initState() {
     super.initState();
-    _songs = MusicService().fetchAllMusicFiles(); // Chargez toutes les musiques
+    _songs = MusicService().fetchAllMusicFiles(); // Chargement de toutes les musiques
   }
 
   @override
