+++ pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   file_picker: ^8.0.6
   audio_query: ^0.0.2
   on_audio_query: ^2.9.0
+  audio_service: ^0.18.15
 
 
 dev_dependencies:
